Modularise code pieces into different files
