+++ src/input-parser/properties/constants/tokens.ts
@@ -1,0 +1,3 @@
+export const NEWLINE_SEPARATOR = "\n";
+
+export const EQUAL_SEPARATOR = "=";
