+++ src/input-parser/properties/errors/properties-parse.error.ts
@@ -1,0 +1,10 @@
+export class PropertiesParseError extends Error {
+  name = "PropertiesParseError";
+  code = "MISSING_VALUE_ASSIGN";
+  message: string;
+
+  constructor(property: string) {
+    super();
+    this.message = `Property '${property}' is not assigned a value`;
+  }
+}
