+++ src/input-parser/properties/properties-validator.ts
@@ -1,34 +1,12 @@
-class PropertiesParseError extends Error {
-  name = "PropertiesParseError";
-  code = "MISSING_VALUE_ASSIGN";
-  message: string;
-
-  constructor(property: string) {
-    super();
-    this.message = `Property '${property}' is not assigned a value`;
-  }
-}
-
-const formatPropertiesParseError = (err: PropertiesParseError): string => {
-  return `${err.name} - caused by ${err.code}`;
-};
-
-const isNonEmptyLine = (line: string): boolean => {
-  return line.trim().length > 0;
-};
-
-const isLineValid = (line: string): boolean => {
-  const EQUAL_SEPARATOR = "=";
-  if (!line.includes(EQUAL_SEPARATOR)) {
-    throw new PropertiesParseError(line);
-  } else {
-    return true;
-  }
-};
+import { formatParseError } from "../utils/error-utils";
+import { PropertiesParseError } from "./errors/properties-parse.error";
+import {
+  isLineValid,
+  splitPropertiesIntoLines,
+} from "./utils/properties-utils";
 
 const parseProperties = (content: string): boolean => {
-  const NEWLINE_SEPARATOR = "\n";
-  const lines = content.split(NEWLINE_SEPARATOR).filter(isNonEmptyLine);
+  const lines = splitPropertiesIntoLines(content);
 
   try {
     lines.forEach(isLineValid);
@@ -36,7 +14,7 @@ const parseProperties = (content: string): boolean => {
   } catch (err) {
     // PropertiesParseError
     if (err instanceof PropertiesParseError) {
-      console.error(formatPropertiesParseError(err));
+      console.error(formatParseError(err));
       // PropertiesParseError - caused by MISSING_VALUE_ASSIGN
       // = Property is not assigned a value
     } else {
@@ -46,6 +24,10 @@ const parseProperties = (content: string): boolean => {
   }
 };
 
+/**
+ * Return true if content is valid properties.
+ * @param content - Properties file content
+ */
 export const validateDotProperties = (content: string): boolean => {
   return parseProperties(content);
 };
