+++ src/input-parser/properties/types/line.type.ts
@@ -1,0 +1,1 @@
+export type Line = string;
