+++ src/input-parser/yaml/utils/yaml-utils.ts
@@ -1,0 +1,9 @@
+import { Document } from "yaml";
+
+export const isDocumentValid = (document: Document): boolean => {
+  if (document.errors.length > 0) {
+    throw document.errors[0];
+  } else {
+    return true;
+  }
+};
